Updated project structure. Added automatic authToken refresh functionality.

+++ BluemixObjectStorage/BluemixObjectStorage/ObjectStorage.swift
@@ -37,7 +37,7 @@ public class ObjectStorage {
 	internal var authTokenManager:AuthTokenManager?
 
 	/**
-	Initialize ObjectStore by supplying projectId and optionally requestManager
+	Initialize ObjectStore by supplying projectId
 
 	- Parameter projectId: ProjectId provided by the IBM Object Store. Can be obtained via VCAP_SERVICES, service instance keys or IBM Object Store dashboard.
 	*/
@@ -47,7 +47,7 @@ public class ObjectStorage {
 	}
 
 	/**
-	Connect to ObjectStorage using userId and password
+	Connect to ObjectStorage using userId and password.
 
 	- Parameter userId: UserId provided by the IBM Object Store. Can be obtained via VCAP_SERVICES, service instance keys or IBM Object Store dashboard.
 	- Parameter password: Password provided by the IBM Object Store. Can be obtained via VCAP_SERVICES, service instance keys or IBM Object Store dashboard.
@@ -56,30 +56,37 @@ public class ObjectStorage {
 	public func connect(userId userId:String, password:String, region:String, completionHandler:(error: ObjectStorageError?) -> Void) {
 		self.authTokenManager = AuthTokenManager(projectId: projectId, userId: userId, password: password)
 		
-		authTokenManager?.refreshAuthToken { (error) in
+		authTokenManager?.refreshAuthToken { (error, authToken) in
 			if error != nil {
+				self.authTokenManager = nil
 				completionHandler(error: ObjectStorageError.FailedToRetrieveAuthToken)
 			} else {
 				self.projectResource = (region == ObjectStorage.REGION_DALLAS) ?
 					ObjectStorage.DALLAS_RESOURCE.resourceByAddingPathComponent(pathComponent: self.projectId) :
 					ObjectStorage.LONDON_RESOURCE.resourceByAddingPathComponent(pathComponent: self.projectId)
+				
 				completionHandler(error: nil)
 			}
 		}
 	}
 	
 	/**
-	Connect to ObjectStorage using preobtained authToken
+	Connect to ObjectStorage using pre-obtained authToken.
 	
 	- Parameter authToken: authToken obtained from Identity Server
 	- Parameter region: Defines whether ObjectStore should connect to Dallas or London instance of IBM Object Store. Use *ObjectStore.REGION_DALLAS* and *ObjectStore.REGION_LONDON* as values
 	*/
 	public func connect(authToken:String, region:String, completionHandler:(error: ObjectStorageError?) -> Void) {
 		self.authTokenManager = AuthTokenManager(projectId: projectId, authToken: authToken)
+		
 		self.retrieveContainersList { (error, containers) in
 			if let error = error{
+				self.authTokenManager = nil
 				completionHandler(error: error)
 			} else {
+				self.projectResource = (region == ObjectStorage.REGION_DALLAS) ?
+					ObjectStorage.DALLAS_RESOURCE.resourceByAddingPathComponent(pathComponent: self.projectId) :
+					ObjectStorage.LONDON_RESOURCE.resourceByAddingPathComponent(pathComponent: self.projectId)
 				completionHandler(error: nil)
 			}
 		}
@@ -90,7 +97,6 @@ public class ObjectStorage {
 	Create a new container
 
 	- Parameter name: The name of container to be created
-	- Parameter completionHandler: Closure to be executed once container is created.
 	*/
 	public func createContainer(name name:String, completionHandler: (error:ObjectStorageError?, container: ObjectStorageContainer?) -> Void){
 		logger.info("Creating container [\(name)]")
@@ -100,18 +106,24 @@ public class ObjectStorage {
 			return completionHandler(error: ObjectStorageError.NotConnected, container: nil)
 		}
 		
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken)
-		let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
-		
-		HttpClient.put(resource: resource!, headers: headers) { error, status, headers, data in
-			if let error = error {
-				completionHandler(error: ObjectStorageError.from(httpError: error), container: nil)
-			} else {
-				self.logger.info("Created container [\(name)]")
-				let container = ObjectStorageContainer(name: name, resource: resource!, objectStore: self)
-				completionHandler(error: nil, container: container)
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error, container: nil)
 			}
-		}
+			
+			let headers = Utils.createHeaderDictionary(authToken: authToken)
+			let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
+			
+			HttpClient.put(resource: resource!, headers: headers) { error, status, headers, data in
+				if let error = error {
+					completionHandler(error: ObjectStorageError.from(httpError: error), container: nil)
+				} else {
+					self.logger.info("Created container [\(name)]")
+					let container = ObjectStorageContainer(name: name, resource: resource!, objectStore: self)
+					completionHandler(error: nil, container: container)
+				}
+			}
+		})
 	}
 
 
@@ -119,7 +131,6 @@ public class ObjectStorage {
 	Retrieve an existing container
 
 	- Parameter name: The name of container to retrieve
-	- Parameter completionHandler: Closure to be executed once container is retrieved.
 	*/
 	public func retrieveContainer(name name:String, completionHandler: (error:ObjectStorageError?, container: ObjectStorageContainer?) -> Void){
 		logger.info("Retrieving container [\(name)]")
@@ -129,24 +140,29 @@ public class ObjectStorage {
 			return completionHandler(error: ObjectStorageError.NotConnected, container: nil)
 		}
 
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken)
-		let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
-		
-		HttpClient.get(resource: resource!, headers: headers) { error, status, headers, data in
-			if let error = error {
-				completionHandler(error: ObjectStorageError.from(httpError: error), container: nil)
-			} else {
-				self.logger.info("Retrieved container [\(name)]")
-				let container = ObjectStorageContainer(name: name, resource: resource!, objectStore: self)
-				completionHandler(error: nil, container: container)
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error, container: nil)
 			}
-		}
+
+			let headers = Utils.createHeaderDictionary(authToken: authToken)
+			let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
+			
+			HttpClient.get(resource: resource!, headers: headers) { error, status, headers, data in
+				if let error = error {
+					completionHandler(error: ObjectStorageError.from(httpError: error), container: nil)
+				} else {
+					self.logger.info("Retrieved container [\(name)]")
+					let container = ObjectStorageContainer(name: name, resource: resource!, objectStore: self)
+					completionHandler(error: nil, container: container)
+				}
+			}
+		})
 	}
 
 	/**
 	Retrieve a list of existing containers
 
-	- Parameter completionHandler: Closure to be executed once list of containeds is retrieved.
 	*/
 	public func retrieveContainersList(completionHandler completionHandler: (error:ObjectStorageError?, containers: [ObjectStorageContainer]?) -> Void){
 		logger.info("Retrieving containers list")
@@ -155,37 +171,43 @@ public class ObjectStorage {
 			logger.error(String(ObjectStorageError.NotConnected))
 			return completionHandler(error: ObjectStorageError.NotConnected, containers: nil)
 		}
-
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken)
 		
-		HttpClient.get(resource: projectResource!, headers: headers) {error, status, headers, data in
-			if let error = error{
-				completionHandler(error: ObjectStorageError.from(httpError: error), containers: nil)
-			} else {
-				self.logger.info("Retrieved containers list")
-				var containersList = [ObjectStorageContainer]()
-				let responseBodyString = String(data: data!, encoding: NSUTF8StringEncoding)!
-				
-				let containerNames = responseBodyString.componentsSeparatedByString("\n")
-				
-				for containerName:String in containerNames{
-					if containerName.characters.count == 0 {
-						continue
-					}
-					let containerResource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + containerName))
-					let container = ObjectStorageContainer(name: containerName, resource: containerResource!, objectStore: self)
-					containersList.append(container)
-				}
-				completionHandler(error: nil, containers: containersList)
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error, containers: nil)
 			}
-		}
+
+			let headers = Utils.createHeaderDictionary(authToken: authToken)
+			
+			HttpClient.get(resource: self.projectResource!, headers: headers) {error, status, headers, data in
+				if let error = error{
+					completionHandler(error: ObjectStorageError.from(httpError: error), containers: nil)
+				} else {
+					self.logger.info("Retrieved containers list")
+					var containersList = [ObjectStorageContainer]()
+					let responseBodyString = String(data: data!, encoding: NSUTF8StringEncoding)!
+					
+					let containerNames = responseBodyString.componentsSeparatedByString("\n")
+					
+					for containerName:String in containerNames{
+						if containerName.characters.count == 0 {
+							continue
+						}
+						let containerResource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + containerName))
+						let container = ObjectStorageContainer(name: containerName, resource: containerResource!, objectStore: self)
+						containersList.append(container)
+					}
+					completionHandler(error: nil, containers: containersList)
+				}
+			}
+		})
+
 	}
 
 	/**
 	Delete an existing container
 
 	- Parameter name: The name of container to delete
-	- Parameter completionHandler: Closure to be executed once container is deleted.
 	*/
 	public func deleteContainer(name name:String, completionHandler: (error:ObjectStorageError?) -> Void){
 		logger.info("Deleting container [\(name)]")
@@ -195,17 +217,23 @@ public class ObjectStorage {
 			return completionHandler(error: ObjectStorageError.NotConnected)
 		}
 
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken)
-		let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
-		
-		HttpClient.delete(resource: resource!, headers: headers) { error, status, headers, data in
-			if let error = error {
-				completionHandler(error: ObjectStorageError.from(httpError: error))
-			} else {
-				self.logger.info("Deleted container [\(name)]")
-				completionHandler(error: nil)
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error)
 			}
-		}
+
+			let headers = Utils.createHeaderDictionary(authToken: authToken)
+			let resource = self.projectResource?.resourceByAddingPathComponent(pathComponent: Utils.urlPathEncode(text: "/" + name))
+			
+			HttpClient.delete(resource: resource!, headers: headers) { error, status, headers, data in
+				if let error = error {
+					completionHandler(error: ObjectStorageError.from(httpError: error))
+				} else {
+					self.logger.info("Deleted container [\(name)]")
+					completionHandler(error: nil)
+				}
+			}
+		})
 	}
 	
 
@@ -213,33 +241,36 @@ public class ObjectStorage {
 	Update account metadata
 
 	- Parameter metadata: a dictionary of metadata items, e.g. ["X-Account-Meta-Subject":"AmericanHistory"]. It is possible to supply multiple metadata items within same invocation. To delete a particular metadata item set it's value to an empty string, e.g. ["X-Account-Meta-Subject":""]. See Object Storage API v1 for more information about possible metadata items - http://developer.openstack.org/api-ref-objectstorage-v1.html
-	- Parameter completionHandler: Closure to be executed once metadata is updated.
 	*/
 	public func updateMetadata(metadata metadata:Dictionary<String, String>, completionHandler: (error:ObjectStorageError?) -> Void){
 		logger.info("Updating metadata :: \(metadata)")
 		
-		guard projectResource != nil else{
+		guard projectResource != nil else {
 			logger.error(String(ObjectStorageError.NotConnected))
 			return completionHandler(error: ObjectStorageError.NotConnected)
 		}
-
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken, additionalHeaders: metadata)
-
-		HttpClient.post(resource: projectResource!, headers: headers) { error, status, headers, data in
-			if let error = error {
-				completionHandler(error:ObjectStorageError.from(httpError: error))
-			} else {
-				self.logger.info("Metadata updated :: \(metadata)")
-				completionHandler(error:nil)
+		
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error)
 			}
-		}
+
+			let headers = Utils.createHeaderDictionary(authToken: authToken, additionalHeaders: metadata)
+			HttpClient.post(resource: self.projectResource!, headers: headers) { error, status, headers, data in
+				if let error = error {
+					completionHandler(error:ObjectStorageError.from(httpError: error))
+				} else {
+					self.logger.info("Metadata updated :: \(metadata)")
+					completionHandler(error:nil)
+				}
+			}
+		});
 
 	}
 
 	/**
 	Retrieve account metadata. The metadata will be returned to a completionHandler as a Dictionary<String, String> instance with set of keys and values
 
-	- Parameter completionHandler: Closure to be executed once metadata is retrieved.
 	*/
 	public func retrieveMetadata(completionHandler completionHandler: (error: ObjectStorageError?, metadata: [String:String]?) -> Void) {
 		logger.info("Retrieving metadata")
@@ -248,15 +279,21 @@ public class ObjectStorage {
 			logger.error(String(ObjectStorageError.NotConnected))
 			return completionHandler(error: ObjectStorageError.NotConnected, metadata: nil)
 		}
-		let headers = Utils.createHeaderDictionary(authToken: authTokenManager?.authToken)
-		HttpClient.head(resource: projectResource!, headers: headers) { error, status, headers, data in
-			if let error = error {
-				completionHandler(error: ObjectStorageError.from(httpError: error), metadata: nil)
-			} else {
-				self.logger.info("Metadata retrieved :: \(headers)")
-				completionHandler(error: nil, metadata: headers);
+		
+		authTokenManager?.refreshAuthToken({ (error, authToken) in
+			guard error == nil else {
+				return completionHandler(error: error, metadata: nil)
 			}
-		}
+			let headers = Utils.createHeaderDictionary(authToken: authToken)
+			HttpClient.head(resource: self.projectResource!, headers: headers) { error, status, headers, data in
+				if let error = error {
+					completionHandler(error: ObjectStorageError.from(httpError: error), metadata: nil)
+				} else {
+					self.logger.info("Metadata retrieved :: \(headers)")
+					completionHandler(error: nil, metadata: headers);
+				}
+			}
+		})
 	}
 }
